# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Salsa de Tomate API
  version: 1.0.0
  description: |
    API REST para la red social de recetas.

    Esta especificación documenta el MVP backend con:
    - autenticación por Bearer token
    - recetas con estado draft/published
    - galería y foto principal obligatoria para publicar
    - ratings y reseñas (usuarios autenticados o invitados)
    - feed de recomendación diaria

    > **Creador: Jose Manuel Castillo Queh**
servers:
  - url: http://localhost:8000
    description: Servidor local
security:
  - bearerAuth: []
...
paths:
...
  /api/recipes/{recipe}:
    get:
      tags: [Recipes]
      summary: Obtener detalle de receta
      security: []
      parameters:
        - $ref: '#/components/parameters/RecipeId'
      responses:
        '200':
          description: Detalle completo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Recipes]
      summary: Actualizar receta
      parameters:
        - $ref: '#/components/parameters/RecipeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeUpdateRequest'
            example:
              summary: Ajuste de sazón
              servings: 6
      responses:
        '200':
          description: Receta actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    delete:
      tags: [Recipes]
      summary: Eliminar receta
      parameters:
        - $ref: '#/components/parameters/RecipeId'
      responses:
        '200':
          description: Receta eliminada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: Receta eliminada correctamente.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  parameters:
    RecipeId:
      in: path
      name: recipe
      required: true
      schema:
        type: integer
  responses:
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
          example:
            message: Unauthenticated.
    Forbidden:
      description: Sin permisos
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
          example:
            message: This action is unauthorized.
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
  schemas:
...
    RecipeCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 255
        summary:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        servings:
          type: integer
          nullable: true
        prep_time_minutes:
          type: integer
          nullable: true
        cook_time_minutes:
          type: integer
          nullable: true
        status:
          type: string
          enum: [draft, published]
          nullable: true
        category_ids:
          type: array
          items:
            type: integer
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/RecipeIngredientInput'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/RecipeStepInput'
        video_type:
          type: string
          enum: [upload, youtube]
          nullable: true
        video_url:
          type: string
          format: uri
          nullable: true
    RecipeUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/RecipeCreateRequest'
    RecipeIngredientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        quantity:
          type: number
          nullable: true
        unit:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        calories:
          type: number
          nullable: true
        protein_g:
          type: number
          nullable: true
        carbs_g:
          type: number
          nullable: true
        fat_g:
          type: number
          nullable: true
    RecipeStepInput:
      type: object
      required: [instruction]
      properties:
        instruction:
          type: string
        image_path:
          type: string
          nullable: true
    RecipeSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        servings:
          type: integer
          nullable: true
        prep_time_minutes:
          type: integer
          nullable: true
        cook_time_minutes:
          type: integer
          nullable: true
        status:
          type: string
          enum: [draft, published]
        published_at:
          type: string
          format: date-time
          nullable: true
        average_rating:
          type: number
          format: float
          nullable: true
        ratings_count:
          type: integer
          nullable: true
        likes_count:
          type: integer
          nullable: true
        comments_count:
          type: integer
          nullable: true
        author:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        main_image:
          $ref: '#/components/schemas/RecipeMedia'
    RecipeDetail:
      allOf:
        - $ref: '#/components/schemas/RecipeSummary'
        - type: object
          properties:
            nutrition:
              type: object
              nullable: true
              properties:
                calories:
                  type: number
                protein_g:
                  type: number
                carbs_g:
                  type: number
                fat_g:
                  type: number
            ingredients:
              type: array
              items:
                type: object
                properties:
                  id: { type: integer }
                  position: { type: integer }
                  name: { type: string }
                  quantity: { type: number, nullable: true }
                  unit: { type: string, nullable: true }
                  notes: { type: string, nullable: true }
            steps:
              type: array
              items:
                type: object
                properties:
                  id: { type: integer }
                  step_number: { type: integer }
                  instruction: { type: string }
                  image_path: { type: string, nullable: true }
            media:
              type: array
              items:
                $ref: '#/components/schemas/RecipeMedia'
    RecipeMedia:
      type: object
      nullable: true
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [image, video]
        url:
          type: string
          nullable: true
        path:
          type: string
          nullable: true
        external_url:
          type: string
          nullable: true
        mime_type:
          type: string
          nullable: true
        size_bytes:
          type: integer
          nullable: true
        is_primary:
          type: boolean
        sort_order:
          type: integer
...
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        is_global: { type: boolean }
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string